Add hackish KolibriUserBehaviour setup method
Current locust.io version doesn't have direct support for setup/teardown
methods (to execute code only once before all the tasks start executing)
so we added kind-of-hackish solution which will have to make due until
this update makes the next stable locust.io release.

The primary idea behind the setup method was to be able to retrieve
Kolibri users only once (rather than doing it within each task).

Simliarly, support for the teardown method could be added, but at this
point we don't have a need for it so we'll avoid adding to-be-refactored
code.